+++ packages/expo-ui/src/swift-ui/index.tsx
@@ -24,6 +24,7 @@ export * from './Spacer';
 export * from './Stepper';
 export * from './Switch';
 export * from './Text';
+export * from './Toggle';
 export * from './TextField';
 export * from './SecureField';
 export * from './Namespace';